BACKLOG-12167: rename project into app-shell
