+++ src/javascript/jahia.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line no-undef, camelcase
-__webpack_public_path__ = `${window.contextJsParameters.contextPath}/modules/dx-commons-webpack/javascript/apps/`;
+__webpack_public_path__ = `${window.contextJsParameters.contextPath}/modules/app-shell/javascript/apps/`;
 
 // Configuration
 import './apollo/register';
