+++ src/main/resources/root.jsp
@@ -20,7 +20,7 @@
     <internal:gwtImport module="empty"/>
 
     <link rel="stylesheet" type="text/css" media="screen" href="<c:url value='/engines/jahia-anthracite/css/edit_V8_en.css'/>" />
-    <link rel="stylesheet" type="text/css" media="screen" href="<c:url value='/modules/dx-commons-webpack/css/loader.css'/>" />
+    <link rel="stylesheet" type="text/css" media="screen" href="<c:url value='/modules/app-shell/css/loader.css'/>" />
 </head>
 
 <body>
@@ -50,8 +50,8 @@
 
 <script>window.anthraciteV8 = true;</script>
 <script type="text/javascript" src="<c:url value='/engines/jahia-anthracite/js/dist/build/anthracite-min.js'/>"></script>
-<script src="${contextPath}/modules/dx-commons-webpack/javascript/commons/jahia-commons.dll.js"></script>
-<script src="${contextPath}/modules/dx-commons-webpack/javascript/apps/appshell.js"></script>
+<script src="${contextPath}/modules/app-shell/javascript/commons/jahia-commons.dll.js"></script>
+<script src="${contextPath}/modules/app-shell/javascript/apps/appshell.js"></script>
 <script type="text/javascript">
     Object.assign(window.contextJsParameters, {
         version: '<%= Jahia.VERSION %>',
